% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centers.R
\name{compute_centers}
\alias{compute_centers}
\title{Compute the Centers of A Time-Series}
\usage{
compute_centers(data, daily_mean_resistor)
}
\arguments{
\item{data}{Formatted data.frame. The wide-formatted data to compute centers for.}

\item{daily_mean_resistor}{Character. The variable name which contains the daily mean resistor readings.}
}
\value{
A list containing the dry and wet centers.
}
\description{
Computes the centers of a time-series using information from the resistor readings. Once the centers have been computed they can then be used for classification.
}
\examples{
# Start by formatting the WWC data
formatted_wwc <- format_data(data = wwc,
                             date_var = "date_time",
                             date_start = "2015-04-01",
                             date_end = "2015-08-30",
                             site_id_var = "site_id",
                             sensor_type_var = "sensor_type",
                             resistor_level = "Resistor",
                             water_temp_level = "Water temp",
                             sensor_reading_var = "sensor_reading")

# Then compute the centers of the wet and dry streams
centers <- compute_centers(data = formatted_wwc,
                           daily_mean_resistor = "daily_mean_resistor")
centers

}
