% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_offset.R
\name{find_best_offset}
\alias{find_best_offset}
\title{Estimating Offsets for Log-Log Stage Discharge Relationships}
\usage{
find_best_offset(stages, discharges, era.name, stream.name, export.dir)
}
\arguments{
\item{stages}{A numeric vector of stage values, associated with discharges.}

\item{discharges}{A numeric vector of discharges, associated with stages.}

\item{era.name}{What time interval?}

\item{stream.name}{What stream?}

\item{export.dir}{Directory to put the graphs into.}
}
\description{
Given paired stage and discharge values, calculate the best stage offset
to create a linear relationship in log-log space.
}
\details{
USGS Water Supply Paper 2175 gives a process for creating stage-discharge
relationships in streams.  Basically, you plot out the measured discharges
and their associated stages in log-log space.  Then you do it again,
but this time you subtract an offset from all the stage values.  You continue
this successive approximation until you get a straight line in log-log space.

This function does that, but faster and better.

Estimates happen at 0.01 intervals from 0 to just below the minimum measured
stage value.
}
\examples{
stages <- c(1.2, 1.5, 1.3, 1.0)
discharges <- c(22, 55, 31, 21)
era.name <- '2008-2012'
stream.name <- 'Demo Stream'
export.dir <- './results/offsets/'
find_best_offset(stages, discharges, era.name, stream.name)
}
\keyword{cfs}
\keyword{discharge}
\keyword{flow}
\keyword{stage}
\keyword{stage-discharge}
\keyword{stream}
