% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for classification}
\usage{
format_data(data, date_var, date_start, date_end, site_id_var,
  sensor_type_var, resistor_level, water_temp_level, sensor_reading_var)
}
\arguments{
\item{data}{Unformatted data object.}

\item{date_var}{Character. The name of the date variable. Should be a POSIX (datetime) variable.}

\item{date_start, date_end}{Character. The start and end dates to use for classification.}

\item{site_id_var}{Character. The name of the variable which contains the site ID information. Site IDs should be unique for each sensor.}

\item{sensor_type_var}{Character. The variable in the data which states the type of sensor, whether it be the resistor readings or the water temperature reading.}

\item{resistor_level}{Character. The name of the resistor sensor in the sensor_type_var argument.}

\item{water_temp_level}{Character. The name of the water temperature sensor in the sensor_type_var argument.}
}
\value{
Returns a wide-format data object which can be used in the classification functions.
}
\description{
Takes raw data and formats it for use in the classification functions.
}
\examples{
# The wwc data is packaged along with the streamflow package.
format_data(data = wwc,
            date_var = "date_time",
            date_start = "2015-04-01",
            date_end = "2015-08-30",
            site_id_var = "site_id",
            sensor_type_var = "sensor_type",
            resistor_level = "Resistor",
            water_temp_level = "Water temp",
            sensor_reading_var = "sensor_reading")

}
