% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamflow.R
\docType{package}
\name{streamflow}
\alias{streamflow}
\alias{streamflow-package}
\title{streamflow: A package for classifying streams as wet or dry}
\description{
The streamflow package includes methods which allow users to classify streams as either wet or dry.
}
